+++ libcoda/coda-sp3.c
@@ -866,7 +866,7 @@ static int get_line(FILE *f, char *line)
         line[0] = '\0';
         return 0;
     }
-    length = strlen(line);
+    length = (long)strlen(line);
 
     /* remove 'linefeed' character if available */
     if (length > 0 && line[length - 1] == '\n')
